Moved most of the logic to a Factory extension
Refs commit: KnpLabs/KnpMenu@5056422affd0d9f027f9f2cb75c2ccdf34964db

Again, this allows us to have less duplication.
