+++ DependencyInjection/CmfMenuExtension.php
@@ -32,8 +32,7 @@ class CmfMenuExtension extends Extension
         );
 
         $loader->load('menu.xml');
-        $factory = $container->getDefinition($this->getAlias().'.factory');
-        $factory->replaceArgument(1, new Reference($config['content_url_generator']));
+        $container->setAlias('cmf_menu.content_router', $config['content_url_generator']);
         $container->setParameter($this->getAlias() . '.allow_empty_items', $config['allow_empty_items']);
 
         $this->loadVoters($config, $loader, $container);
